requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.13.0
fpdf2>=2.7.5
python-pptx>=0.6.21
pillow>=10.0.0
python-dateutil>=2.8.2
scipy>=1.11.0
statsmodels>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
pulp>=2.7.0
streamlit-agraph>=0.0.45 